+++ .vscode/settings.json
@@ -1,3 +1,8 @@
 {
-  "cSpell.words": ["Dootstrapper", "bootstrapper"]
+  "cSpell.words": [
+    "Codebuild",
+    "Codepipeline",
+    "Dootstrapper",
+    "bootstrapper"
+  ]
 }
